requirements.txt
# Core dependencies
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0

# Sandbox execution
e2b-code-interpreter>=1.5.2

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Type checking and code quality
mypy>=1.7.0
black>=23.12.0
ruff>=0.1.8

# Scientific computing (for sandbox environment)
numpy>=1.24.0
sympy>=1.12
pint>=0.23.0

# Image processing (for Active Vision CV tools)
Pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.11.0

# Benchmarking
datasets>=2.14.0          # HuggingFace datasets
pyyaml>=6.0               # Config file support
tqdm>=4.66.0              # Progress bars
